Beta1 Update. The plug-in is now linked to Houdini 18.0.597 / HAPI 3.4.0.
-------------------------------------------------------
New Features:
-------------------------------------------------------


Blueprint Support (experimental):

It is now possible to use Houdini Asset Components in the Blueprint Editor.
This lets you preset and use Houdini Digital Assets in Blueprint Actors.
Changing parameters on the Houdini Asset will automatically update all placed instances of that Blueprint.

After adding the Houdini Asset Component, the Houdini Asset it uses can be selected in the details panel.
This will cook the HDA, and display its parameters interface.

Currently, only Static mesh and Instancer outputs are supported, and only geometry inputs are available when using HDAs in the BP Editor.



PDG Async Import (experimental):

The PDG Asset Link now has the ability to use an asynchronous importer commandlet to import WorkItem outputs once they are ready.

The commandlet communicates with the PDG Asset Link via a message bus, and automatically imports the BGEO files from output work items, creating uassets.
After a successful import, the commandlet notifies the PDG asset link which will then instantiate the uasset in the level.

Asynchronous import can be enabled/started/stopped via the "Work Item Import Settings" entry in the Houdini Engine menu.
The PDG Asset Link will display the status of the async importer in its details UI when using it.



BGEO Import Commandlet (experimental):

The commandlet that is used for Asynchronous import of a PDG Graph's work item outputs can also be used independently, to automatically import BGEO files to an unreal project.

The -watch=<path> command-line argument can be used to place the commandlet into directory watching mode.

The BGEO import commandlet supports importing to bake path, by reading the "unreal_bake_folder" and "unreal_output_name" name attributes when in "bake" mode.

- To watch a directory for new files and import them as uassets to the HoudiniEngine temporary path: 
    <editor_exe> <uproject file path or project name> -run=HoudiniGeoImport -watch=<path>
- To watch a directory for new files and import them as uassets to bake paths (respecting "unreal_bake_folder" and "unreal_output_name"): 
    <editor_exe> <uproject file path or project name> -run=HoudiniGeoImport -bake -watch=<path>
- To import a single .bgeo file as a uasset to the HoudiniEngine temporary path: 
    <editor_exe> <uproject file path or project name> -run=HoudiniGeoImport <path to bgeo file>
- To import a single .bgeo file as a uasset to the bake path (respecting "unreal_bake_folder" and "unreal_output_name"): 
    <editor_exe> <uproject file path or project name> -run=HoudiniGeoImport -bake <path to bgeo file>
- To display/log the help:
    <editor_exe> <uproject file path or project name> -run=HoudiniGeoImport -help



Curve Output:

It is now possible to output Spline Component from an HDA.
To do so, curves in the HDA need to be marked using the "unreal_output_curve" primitive attribute (value set to 1).
You can also control if you want the spline to be linear/closed using the "unreal_output_curve_linear" and "unreal_output_curve_closed" primitive attributes respectively.



Session Sync - Viewport Sync:

When using Session Sync, you can decide to link the Houdini and Unreal viewport via the "Sync Viewport" entry in the Houdini Engine menu.
You can decide to sync both viewport, or have unreal control the houdini viewport, or the opposite.
Viewport Sync currently only syncs the cameras' positions and orientations, and not their FOV, so the two views will not be 100% identical.



Backward compatibility with Version1:

The plugin is now backward compatible with version 1.
When loading a level that contains Houdini objects made with version 1, the plugin now tries to convert the V1 components, parameters, inputs and outputs to their v2 equivalents.

Some HDAs might need to be rebuilt after the conversion for their parameters and inputs to be displayed properly by the v2 plugin.
Certain types of parameters or inputs are currently not properly upgraded to v2 (for example Editable curves or Handles).

The conversion of the legacy data is still in progress and will be improved upon in the future.

The legacy conversion can be enabled or disabled in the plugin settings.
If disabled, legacy objects will simply be ignored when loading the level.
The legacy settings also have an option to automatically rebuild converted Houdini Asset Components.

This was added to facilitate migrating a version 1 project to version 2.
To do so, we highly recommend duplicating (or at least backing up) your project first.
The version 1 plugin should then be deleted from the Plugins/Runtime folder before copying the version 2 files. 
The conversion of the legacy objects will be done automatically upon loading the levels.



-------------------------------------------------------
Bug Fixes:
-------------------------------------------------------


- Generic UProperty Attributes now properly handle nested UStructs and Array properties.
  This fixes crashes when attempting to edit a property stored in nested struct or arrays via a generic property attribute (unreal_uproperty_)
  Attribute tuples can now successfully be used to set array UProperties.
- Fixed Scale Y/Z values being swapped (Issue #44)
- Fixed PackBeforeMerge, KeepWorldTransform being applied only to the first input mesh found in a BP or Actor. (Geo/World Input) (Issue #46)
- Fixed crash when importing some StaticMeshes (Issue #47)
- Changing Mobility, Visibility, or Hidden in Game on a HAC is now properly propagated to child components.
- Auto-save is disabled when HDAs are being cooked or are processing outputs.
- Fixed baking ProxyMeshes.
  When baking a HoudiniAssetComponent that uses Proxy Meshes, first refine the meshes to StaticMesh before baking.
- Fixed Parameter Details UI bug when cooking is paused.
- Fixed Ramp Parameter Details showing tooltips inside the curve editor.
- Fixed crash when using FMeshDescription to output meshes.
- We're using SleepNoStats() instead of Sleep() in the Scheduler, this avoids stat messages accumulating and allocating too much memory when the editor hangs.
- Fixed landscape input's "mesh" and "points" modes not functioning/creating anything on the Houdini side.
- Fixed FHoudiniEngineBakeUtils::CopyActorContentsToBlueprint returning nullptr instead of false.
- The PDG Asset Link now handles instanced foliage destruction correctly.
- Fixed bug causing incorrect work item numbers being displayed in the PDG Asset Link UI.
- Improved PDG Asset Link UI updates.
- Added undo/redo support to PDG asset link UI.
- Fixed various Linux/Clang build issues.
- Fixed Actors baked by the PDG Asset Link being destroyed/disappearing after reloading/restarting the engine.
- Fixed crash when a PDG work result actor name is already in use.
- PDG asset link: Fixed work item counts and node states not being reset upon Rebuild.